+++ app/src/main/java/com/example/cryptofunding/data/Task.kt
@@ -1,4 +1,4 @@
 package com.example.cryptofunding.data
 
-data class Task constructor(val title: String, val summary: String, val amount: Int, val limitDate: String, var expanded: Boolean = false) {
+data class Task constructor(val title: String, val summary: String, val amount: Double, val limitDate: String, var expanded: Boolean = false) {
 }
